Make cloudflare crate usable in wasm32 targets
Using the cloudflare in a wasm32 target was failing to compile.
The first reason being that the blocking client cannot ever work
in wasm.
The second reason being that the asynchronous client was using
the async_trait that did not work fine under wasm_bindgen either.

As a result, we now hide away those portions of the framework
implementation when targeting wasm32. We do so by resorting to
conditional compilation macros.

To test this out we used the framework to call some endpoints in
a Cloudflare Worker written in Rust (thus running in wasm) and it
all worked fine after these changes.

This thus enables Workers' developers in Rust to use this library
at will and contribute further to it.
