+++ README.md
@@ -7,3 +7,7 @@
 ⚠️ This library is a Work in Progress! ⚠️
 
 This library provides convenience functions that wrap the Cloudflare API.
+
+It provides two client implementations (asynchronous vs blocking). However, 
+projects targeting `wasm32` only get the asynchronous one (as it does not make
+sense to block in that target).
