+++ cloudflare/Cargo.toml
@@ -13,6 +13,7 @@ anyhow = "1.0"
 async-trait = "0.1"
 base64 = "0.13"
 chrono = { version = "0.4", features = ["serde"] }
+cfg-if = "0.1"
 http = "0.2"
 percent-encoding = "1.0"
 reqwest = { version = "0.10", features = ["json", "blocking"] }
