+++ cloudflare/src/framework/async_api.rs
@@ -10,6 +10,7 @@ use crate::framework::{
 use async_trait::async_trait;
 use reqwest;
 use serde::Serialize;
+use cfg_if::cfg_if;
 
 #[async_trait]
 pub trait ApiClient {
@@ -45,10 +46,18 @@ impl Client {
         config: HttpApiClientConfig,
         environment: Environment,
     ) -> anyhow::Result<Client> {
-        let http_client = reqwest::Client::builder()
-            .timeout(config.http_timeout)
-            .default_headers(config.default_headers)
-            .build()?;
+        #[allow(unused_mut)]
+        let mut builder = reqwest::Client::builder()
+            .default_headers(config.default_headers);
+
+        cfg_if! {
+            // There are no timeouts in wasm.
+            if #[cfg(not(target_arch = "wasm32"))] {
+                builder = builder.timeout(config.http_timeout);
+            }
+        }
+
+        let http_client = builder.build()?;
 
         Ok(Client {
             environment,
@@ -56,11 +65,8 @@ impl Client {
             http_client,
         })
     }
-}
 
-#[async_trait]
-impl ApiClient for Client {
-    async fn request<ResultType, QueryType, BodyType>(
+    pub async fn request_handle<ResultType, QueryType, BodyType>(
         &self,
         endpoint: &(dyn Endpoint<ResultType, QueryType, BodyType> + Send + Sync),
     ) -> ApiResponse<ResultType>
@@ -86,6 +92,23 @@ impl ApiClient for Client {
         request = request.auth(&self.credentials);
         let response = request.send().await?;
         map_api_response(response).await
+    }
+}
+
+// The async_trait does not work nicely in wasm.
+#[cfg(not(target_arch = "wasm32"))]
+#[async_trait]
+impl ApiClient for Client {
+    async fn request<ResultType, QueryType, BodyType>(
+        &self,
+        endpoint: &(dyn Endpoint<ResultType, QueryType, BodyType> + Send + Sync),
+    ) -> ApiResponse<ResultType>
+        where
+            ResultType: ApiResult,
+            QueryType: Serialize,
+            BodyType: Serialize,
+    {
+        self.request_handle(endpoint).await
     }
 }
