+++ cloudflare/src/framework/auth.rs
@@ -1,5 +1,3 @@
-use reqwest::blocking::RequestBuilder;
-
 #[derive(Debug)]
 pub enum Credentials {
     UserAuthKey { email: String, key: String },
@@ -23,13 +21,4 @@ impl Credentials {
 
 pub trait AuthClient {
     fn auth(self, credentials: &Credentials) -> Self;
-}
-
-impl AuthClient for RequestBuilder {
-    fn auth(mut self, credentials: &Credentials) -> Self {
-        for (k, v) in credentials.headers() {
-            self = self.header(k, v);
-        }
-        self
-    }
 }
