+++ cloudflare/src/framework/mod.rs
@@ -4,14 +4,15 @@ This module controls how requests are sent to Cloudflare's API, and how response
 pub mod apiclient;
 pub mod async_api;
 pub mod auth;
+#[cfg(not(target_arch = "wasm32"))]  // There is no blocking implementation for wasm.
+pub mod blocking_api;
 pub mod endpoint;
 pub mod json_utils;
+#[cfg(not(target_arch = "wasm32"))]  // The mock contains a blocking implementation.
 pub mod mock;
 mod reqwest_adaptors;
 pub mod response;
 
-use crate::framework::{apiclient::ApiClient, auth::AuthClient, response::map_api_response};
-use reqwest_adaptors::match_reqwest_method;
 use serde::Serialize;
 use std::time::Duration;
 
@@ -51,6 +52,8 @@ impl<'a> From<&'a Environment> for url::Url {
     }
 }
 
+// There is no blocking support for wasm.
+#[cfg(not(target_arch = "wasm32"))]
 /// Synchronous Cloudflare API client.
 pub struct HttpApiClient {
     environment: Environment,
@@ -72,59 +75,5 @@ impl Default for HttpApiClientConfig {
             http_timeout: Duration::from_secs(30),
             default_headers: http::HeaderMap::default(),
         }
-    }
-}
-
-impl HttpApiClient {
-    pub fn new(
-        credentials: auth::Credentials,
-        config: HttpApiClientConfig,
-        environment: Environment,
-    ) -> anyhow::Result<HttpApiClient> {
-        let http_client = reqwest::blocking::Client::builder()
-            .timeout(config.http_timeout)
-            .default_headers(config.default_headers)
-            .build()?;
-
-        Ok(HttpApiClient {
-            environment,
-            credentials,
-            http_client,
-        })
-    }
-}
-
-// TODO: This should probably just implement request for the Reqwest client itself :)
-// TODO: It should also probably be called `ReqwestApiClient` rather than `HttpApiClient`.
-impl<'a> ApiClient for HttpApiClient {
-    /// Synchronously send a request to the Cloudflare API.
-    fn request<ResultType, QueryType, BodyType>(
-        &self,
-        endpoint: &dyn endpoint::Endpoint<ResultType, QueryType, BodyType>,
-    ) -> response::ApiResponse<ResultType>
-    where
-        ResultType: response::ApiResult,
-        QueryType: Serialize,
-        BodyType: Serialize,
-    {
-        // Build the request
-        let mut request = self
-            .http_client
-            .request(
-                match_reqwest_method(endpoint.method()),
-                endpoint.url(&self.environment),
-            )
-            .query(&endpoint.query());
-
-        if let Some(body) = endpoint.body() {
-            request = request.body(serde_json::to_string(&body).unwrap());
-            request = request.header(reqwest::header::CONTENT_TYPE, endpoint.content_type());
-        }
-
-        request = request.auth(&self.credentials);
-
-        let response = request.send()?;
-
-        map_api_response(response)
     }
 }
